# docs/api/openapi.yaml
openapi: 3.1.0
info: { title: ECsite application API, version: 1.0.0, license: { name: Proprietary, identifier: LicenseRef-Proprietary }, description: 'Detailed design §6. Browser-facing operations; Server Component reads use the data layer directly and are not API paths.' }
servers: [{ url: /api }]
security: []
paths:
...
  /cart:
    get:
      operationId: get_cart
      summary: Read current-price cart for its anonymous cookie or member session
      description: Anonymous carts use an HttpOnly, Secure-in-production, SameSite=Lax signed cookie. A heavy-item shipping amount remains null until its destination is known.
      security: [{ sessionCookies: [] }, {}]
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
    put:
      operationId: put_cart
      summary: Set cart item quantity (1–10)
      security: [{ sessionCookies: [] }, {}]
      x-origin-check: required
      parameters: [{ $ref: '#/components/parameters/Origin' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CartItemInput' } } } }
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/CartConflict' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      operationId: delete_cart_item
      summary: Remove one item from the caller's cart
      security: [{ sessionCookies: [] }, {}]
      x-origin-check: required
      parameters:
        - { $ref: '#/components/parameters/Origin' }
        - { name: productId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    sessionCookies:
      type: apiKey
      in: cookie
      name: Cookie
      description: 'Server-managed Supabase SSR session and/or anonymous cart cookie; cookie names are implementation-owned. HttpOnly, Secure, SameSite=Lax.'
    standardWebhookId: { type: apiKey, in: header, name: webhook-id }
    standardWebhookTimestamp: { type: apiKey, in: header, name: webhook-timestamp }
    standardWebhookSignature: { type: apiKey, in: header, name: webhook-signature }
    internalJobHmac:
      type: apiKey
      in: header
      name: X-Internal-Job-Signature
      description: HMAC-SHA256; see reconcile-payments operation for canonical bytes and timestamp contract. Uses environment-specific INTERNAL_JOB_SECRET.
  parameters:
...
    Origin: { name: Origin, in: header, required: true, schema: { type: string, format: uri } }
...
  schemas:
    Envelope:
      type: object
      required: [data, requestId]
      additionalProperties: false
      properties: { data: {}, requestId: { type: string, minLength: 1 } }
    ErrorEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fieldErrors: { type: object, additionalProperties: { type: array, items: { type: string } } }
        requestId: { type: string, minLength: 1 }
...
    CartProjection:
      type: object
      required: [items, goodsTotalYen, estimatedShippingYen]
      additionalProperties: false
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CartLine' } }
        goodsTotalYen: { type: integer, minimum: 0 }
        estimatedShippingYen: { type: [integer, 'null'], minimum: 0, description: Null while a heavy item's destination-specific official rate cannot be determined. }
    CartLine:
      type: object
      required: [productId, quantity, unitPriceYen, lineTotalYen, availableQuantity]
      properties:
        productId: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 10 }
        unitPriceYen: { type: integer, minimum: 0 }
        lineTotalYen: { type: integer, minimum: 0 }
        availableQuantity: { type: integer, minimum: 0 }
    CartItemInput:
      type: object
      required: [productId, quantity]
      additionalProperties: false
      properties: { productId: { type: string, format: uuid }, quantity: { type: integer, minimum: 1, maximum: 10 } }
...
    CartConflictEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, currentCart]
          properties:
            code: { type: string, const: CONFLICT }
            message: { type: string }
            currentCart: { $ref: '#/components/schemas/CartProjection' }
        requestId: { type: string, minLength: 1 }
...
  responses:
...
    Cart:
      description: Current server-priced cart and estimated shipping.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/Envelope' }
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/CartProjection' }
...
    CartConflict:
      description: Requested quantity exceeded current available stock; current server cart and availability are returned.
      content: { application/json: { schema: { $ref: '#/components/schemas/CartConflictEnvelope' } } }
...
    BadRequest: { description: Invalid input or webhook signature., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Caller lacks required role., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: 'Resource is absent, unpublished or not owned.', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Unavailable: { description: Retryable service failure or unavailable shipping quote; do not guess amount., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }